Search is now a view mixin
