+++ inboxen/templates/inboxen/styleguide.html
@@ -69,15 +69,6 @@
                             <li><a>Item 2</a></li>
                             <li><a>Item 3</a></li>
                         </ul>
-                        <form id="inboxen-search-box" class="navbar-form navbar-right">
-                            <div class="input-group">
-                                <label class="sr-only">Search emails and inboxes</label>
-                                <input type="text" class="form-control" name="q">
-                                <span class="input-group-btn">
-                                    <button class="btn btn-default">Search</button>
-                                </span>
-                            </div>
-                        </form>
                     </div>
                 </div>
             </div>
