+++ inboxen/views/user/__init__.py
@@ -15,4 +15,3 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ##
 from .home import *  # noqa
-from .search import *  # noqa
